% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_trends.R
\name{artist_trends}
\alias{artist_trends}
\title{Plot top artists over time}
\usage{
artist_trends(df, num = 3)
}
\arguments{
\item{df}{Cleaned dataframe (using myspotify::read_file()) of Spotify listening history}

\item{num}{Number of top artists to plot per year}
}
\value{
ggplot2 plot of the top artists listened to over the years
}
\description{
Plot top artists over time
}
\examples{
artist_trends(myspotify::extended_example, num = 5)

}
