% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNE_in.R
\name{DNE_in}
\alias{DNE_in}
\title{List matching -- return items that don't exist}
\usage{
DNE_in(baseline, checkfor)
}
\arguments{
\item{baseline}{List of items to compare against}

\item{checkfor}{List of items to check for in baseline}
}
\value{
Returns list of items in "checkfor" that do NOT exist in "baseline"
}
\description{
List matching -- return items that don't exist
}
\examples{

checking <- c("apple", "banana", "clementine")
for <- c("apple", "peach", "banana")
DNE_in(checking, for)
#> returns "peach"

}
