% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_reason.R
\name{end_reason}
\alias{end_reason}
\title{Most Skipped Songs of All Time}
\usage{
end_reason(df)
}
\value{
text displaying the top three skipped songs
}
\description{
Most Skipped Songs of All Time
}
\examples{
end_reason(extended_example)
}
