% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{extended_example}
\alias{extended_example}
\title{Example of personal data (extended version) downloaded from Spotify}
\format{
\subsection{\code{extended_example}}{

A data frame with 10,835 rows and 9 columns:
\describe{
\item{ts}{POSIXct class time stamp (date and time in Coordinated Universal Time) of song end}
\item{time}{POSIXct class time stamp in Coordinated Universal Time of song end,
derived from datetime with arbitrary Y-M-D for when only time is of interest}
\item{ms_played}{Number of milliseconds played of the song}
\item{song}{Name of the song played}
\item{artist}{Artist of the song played}
\item{album}{Album that the song belongs to}
\item{reason_start, reason_end}{Reasons for starting and ending of a certain song}
...
}
}
}
\source{
Downloaded from personal account Privacy Settings on \url{https://spotify.com}
}
\usage{
extended_example
}
\description{
Listening history data from Spotify including song information
and interaction with given songs
}
\keyword{datasets}
