% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_length.R
\name{filter_length}
\alias{filter_length}
\title{Keep songs based on length of play time}
\usage{
filter_length(df, length = NULL)
}
\arguments{
\item{df}{Dataframe of Spotify listening history (must have \code{ms_played} column)}

\item{length}{The minimum number of minutes the song has been played for}
}
\value{
Dataframe containing only entries that were played for more than or equal to the number of minutes specified in \code{length}
}
\description{
Keep songs based on length of play time
}
\examples{
data <- myspotify$extended_example

two_min <- data \%>\% filter_length(length = 2)

two_half_min <- filter_length(data, 2.5)

}
