% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_playtime.R
\name{filter_playtime}
\alias{filter_playtime}
\title{Keep songs based on length of play time}
\usage{
filter_playtime(df, length = NULL)
}
\arguments{
\item{df}{Dataframe of Spotify listening history (must have \code{ms_played} column)}

\item{length}{The minimum number of minutes the song has been played for}
}
\value{
Dataframe containing only entries that were played for more than or equal to the number of minutes specified in \code{length}
}
\description{
Keep songs based on length of play time
}
\examples{
data <- myspotify::extended_example

two_min <- data \%>\% filter_playtime(length = 2)

two_half_min <- filter_playtime(data, 2.5)

}
