% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topartists.R
\name{most_played_artist}
\alias{most_played_artist}
\title{Top 10 Played Artists and Bands}
\usage{
most_played_artist(extended_data)
}
\arguments{
\item{extended_data}{a data frame}
}
\value{
text and ggplot2 bar graph
}
\description{
Top 10 Played Artists and Bands
}
\examples{
most_played_artist(extended_example)

}
