% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_played_artists.R
\name{most_played_artists}
\alias{most_played_artists}
\title{Top 10 Played Artists and Bands}
\usage{
most_played_artists(df)
}
\arguments{
\item{df}{Cleaned dataframe (using myspotify::read_file()) of Spotify listening history}
}
\value{
text displaying the top three artists and ggplot2 bar chart showing the top 10 listened artists in descending order
}
\description{
Top 10 Played Artists and Bands
}
\examples{
most_played_artists(myspotify::extended_example)
}
