% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_skipped.R
\name{most_skipped}
\alias{most_skipped}
\title{Assessing number of skips by song feature}
\usage{
most_skipped(df, by = "song")
}
\arguments{
\item{df}{Cleaned dataframe (using myspotify::read_file) of Spotity listening history}

\item{by}{The feature by which to check for skips (e.g., song, artist)}
}
\value{
Dataframe of the most skipped songs and the number of skips per song
}
\description{
Assessing number of skips by song feature
}
\examples{
most_skipped(myspotify::extended_example)
most_skipped(myspotify::extended_example, by = "artist")
}
