% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_artists.R
\name{plot_dates}
\alias{plot_dates}
\title{Plot top artists over time}
\usage{
plot_dates(num_artists)
}
\arguments{
\item{num_artists}{The number of artists to show plots for}
}
\description{
Plot top artists over time
}
