% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_times.R
\name{plot_times}
\alias{plot_times}
\title{Plot listening habits by time of day}
\usage{
plot_times(df, targetTZ = "UTC", years = NULL)
}
\arguments{
\item{df}{Data frame of personal Spotify data}

\item{targetTZ}{Desired timezone (default is UTC)}

\item{years}{Year(s) whose data are of interest}
}
\value{
ggplot2 bar chart of listening habits throughout the 24-hour day
}
\description{
Plot listening habits by time of day
}
\examples{
df <- myspotify::extended_example

plot_times(df, targetTZ = "EST", years = c(2019, 2022))

plot_times(df, targetTZ = "Africa/Tripoli", years = 2020) +
   ggplot2::labs(subtitle = "from 2019")
}
