% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read-in Spotify JSON data file}
\usage{
read_file(filepath)
}
\arguments{
\item{filepath}{File path string to where the endsong.json (or endsong_X.json) file is hosted}
}
\value{
dataframe object with select columns
}
\description{
Read-in Spotify JSON data file
}
\examples{
read_file("C:/Users/SophiaSmith/Desktop/my_spotify_data/MyData/endsong_4.json")

}
