% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{readfile}
\alias{readfile}
\title{Read-in Spotify JSON data file}
\usage{
readfile(filepath)
}
\arguments{
\item{filepath}{File path string to where the endsong.json file is hosted}
}
\value{
dataframe object with select columns
}
\description{
Read-in Spotify JSON data file
}
\examples{
read_file()
}
