% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_word.R
\name{search_word}
\alias{search_word}
\title{Search specific word or string in song name}
\usage{
search_word(df, string)
}
\arguments{
\item{string}{}
}
\value{
dataframe of a list of all the songs a user streamed with that specific string in the song name
}
\description{
Search specific word or string in song name
}
\examples{
search_word(extended_example, love)
}
