% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries}
\alias{timeseries}
\title{Plot number of listens over time}
\usage{
timeseries(df, interval = "month", plottype = "bar")
}
\arguments{
\item{df}{Cleaned dataframe (using myspotify::read_file()) of Spotify listening history}

\item{plottype}{Type of plot to produce: can be bar (default) or line}
}
\value{
ggplot2 plot of listening habits in songs per day over
}
\description{
Plot number of listens over time
}
