% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topartists.R
\name{most_played_artist}
\alias{most_played_artist}
\title{Top 10 Played Artists and Bands}
\usage{
most_played_artist(df)
}
\arguments{
\item{df}{Data frame of personal Spotify data}
}
\value{
text displaying the top three artists and ggplot2 bar chart showing the top 10 listened artists in descending order
}
\description{
Top 10 Played Artists and Bands
}
\examples{
most_played_artist(extended_example)
}
