% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_btn.R
\name{play_btn}
\alias{play_btn}
\title{Which Songs You Manually Pressed "Play Button" the Most}
\usage{
play_btn(df)
}
\arguments{
\item{df}{Cleaned dataframe (using myspotify::read_file()) of Spotify listening history}
}
\value{
Dataframe of all the songs the user hit the "play button" to stream
}
\description{
Which Songs You Manually Pressed "Play Button" the Most
}
\examples{
play_btn(myspotify::extended_example)
}
