% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dates.R, R/plot_top_artists.R
\name{plot_dates}
\alias{plot_dates}
\title{Plot number of listens per day over time}
\usage{
plot_dates(num_artists)

plot_dates(num_artists)
}
\arguments{
\item{num_artists}{The number of artists to show plots for}

\item{plottype}{}
}
\description{
Plot number of listens per day over time

Plot top artists over time
}
