% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read-in Spotify JSON data files}
\usage{
read_file(...)
}
\arguments{
\item{...}{Links or file path strings to where the endsong.json (or endsong_X.json) file is hosted, separated by commmas}
}
\value{
dataframe object with select columns
}
\description{
Read-in Spotify JSON data files
}
\examples{
read_file("https://bit.ly/github_json_file")
}
