% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries}
\alias{timeseries}
\title{Plot number of listens over time}
\usage{
timeseries(
  df,
  interval = "month",
  limits = c(min(df$date), max(df$date)),
  breaks = "1 year",
  geom = "bar"
)
}
\arguments{
\item{df}{Cleaned dataframe (using myspotify::read_file()) of Spotify listening history}

\item{interval}{Time interval to round dates to ('day', 'month', or 'year')}

\item{limits}{Date-time limits for the x-axis}

\item{breaks}{Time interval to set x-axis label breaks at}

\item{geom}{Type of plot to produce: can be bar (default) or line}
}
\value{
ggplot2 plot of listening habits in songs per day over
}
\description{
Plot number of listens over time
}
\examples{
df <- myspotify::extended_example
t <- c("2021-01-01 00:00:00", "2022-01-01 00:00:00")
timeseries(df, interval = "month", limits = t, breaks = "3 months")

timeseries(df, interval = "day", breaks = "3 months", geom = "line") +
   ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 270, vjust = 0.5, hjust=1))


}
